working on a wysiwyg editor for adding posts. i started with editorjs but then switch to react-simple-wysiwyg, i like the simplicity. just need to work on getting the data to integrate with react-hook-form.

+++ client/src/utils/Fetch.js
@@ -25,6 +25,10 @@ export const fetchPostById = async (id) => {
 };
 
 export const deletePost = async (id, postList) => {
+  const assure = confirm("Delete this post?");
+  if (!assure) {
+    return;
+  }
   try {
     const response = await fetch(`/api/delete/${id}`, {
       method: "DELETE",
